Made it Compile + lots of updates
fixed crashes in TeamQueuePage, GameSeperator, and EndOfGamePage. Fixed
more issues in TeamQueuePage and it should almost be done. Started work
on public group chats which I will finish when I get back home, as well
as spectate option when right clicking friend. Made chat work in
endofgamepage and a few GUI updates. Added the Garena stuff just so it
can compile - no idea if it's right or not, just want to get it in a
compiling state