Added Spectate Option + More work on GroupChat
Added basic spectate option - need to figure out how to get spectator
data of someone who is spectating someone (only works if you try to
spectate the person actually in the game atm). Added a little image so
the button isn't ugly and added a feature to clear the notification grid
and did a little more work on the Public Chat, should finish it next
commit.