Finish Group Chat
Group chat should be done, just need to add features to the
GroupChatPlayer and it needs to be cleaned up, inheritance will probably
need to be added later but it's not to bad, just lots of Linq. Getting
ready to fix up the endofgamepage to add stats + report / commend
button.