Finished Reconnection...
Spent hours trying to figure out 3 lines of code... spent forever trying
to read messages to find how you get reconnect info... then found it in
the login packet... then the reconnectinfo was an object so I tried to
reconstruct it based on what was needed... then found it was simply a
PlayerCredentialsDto. I just spent 8 hours on figuring out 3 lines. I've
never felt like more of a failure lol. At least it works... right?...

Fixed up InGame while I was at it as you can return to the page so the
overlay wasn't needed anymore.