Fixed(?) Ranked + QueuePopOverlay
Not finished, but doesn't crash anymore. The enemy team gets extremely
messed up until all bans happen, but you can get all the way through the
lobby without crashing so that's a plus. Fixed queuepopoverlay sometimes
being unresponsive / lying about current ranking or saying unranked, and
made queuepopoverlay not resize so it looks good when you move it
around, and removed the Unlocked image as it was to similar to the
locked. Need to edit the locked image, just doesn't look good with the
rest of the client.