Fix queues + revert Launch
Believe I finally fixed queues? Disabled the solo queue button until I
implement it there. Moved logic out of QueuePopOverlay as that was
causing a lot of issues. Reverted the Client.Launch, having it in ingame
forced it to launch if you launch client while ingame + spammed the
client (never figured out why but if you need it in ingame feel free to
put it back and I can look into why)